+++ jackson/src/test/java/feign/jackson/JacksonCodecTest.java
@@ -1,3 +1,16 @@
+/**
+ * Copyright 2012-2018 The Feign Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package feign.jackson;
 
 import com.fasterxml.jackson.core.JsonGenerator;
@@ -13,10 +26,13 @@ import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 
 import org.junit.Test;
 
+import java.io.Closeable;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -29,6 +45,7 @@ import static feign.Util.UTF_8;
 import static feign.assertj.FeignAssertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 
 public class JacksonCodecTest {
 
@@ -153,6 +170,52 @@ public class JacksonCodecTest {
                                  + "} ]");
   }
 
+  @Test
+  public void decodesIterator() throws Exception {
+    List<Zone> zones = new LinkedList<Zone>();
+    zones.add(new Zone("denominator.io."));
+    zones.add(new Zone("denominator.io.", "ABCD"));
+
+    Response response = Response.builder()
+            .status(200)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .body(zonesJson, UTF_8)
+            .build();
+    Object decoded = JacksonIteratorDecoder.create().decode(response, new TypeReference<Iterator<Zone>>() {}.getType());
+    assertTrue(Iterator.class.isAssignableFrom(decoded.getClass()));
+    assertTrue(Closeable.class.isAssignableFrom(decoded.getClass()));
+    assertEquals(zones, asList((Iterator<?>) decoded));
+  }
+
+  private <T> List<T> asList(Iterator<T> iter) {
+    final List<T> copy = new ArrayList<T>();
+    while (iter.hasNext())
+      copy.add(iter.next());
+    return copy;
+  }
+
+  @Test
+  public void nullBodyDecodesToNullIterator() throws Exception {
+    Response response = Response.builder()
+            .status(204)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .build();
+    assertNull(JacksonIteratorDecoder.create().decode(response, Iterator.class));
+  }
+
+  @Test
+  public void emptyBodyDecodesToNullIterator() throws Exception {
+    Response response = Response.builder()
+            .status(204)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .body(new byte[0])
+            .build();
+    assertNull(JacksonIteratorDecoder.create().decode(response, Iterator.class));
+  }
+
   static class Zone extends LinkedHashMap<String, Object> {
 
     private static final long serialVersionUID = 1L;
@@ -221,5 +284,16 @@ public class JacksonCodecTest {
             .headers(Collections.<String, Collection<String>>emptyMap())
             .build();
     assertThat((byte[]) new JacksonDecoder().decode(response, byte[].class)).isEmpty();
+  }
+
+  /** Enabled via {@link feign.Feign.Builder#decode404()} */
+  @Test
+  public void notFoundDecodesToEmptyIterator() throws Exception {
+    Response response = Response.builder()
+        .status(404)
+        .reason("NOT FOUND")
+        .headers(Collections.<String, Collection<String>>emptyMap())
+        .build();
+    assertThat((byte[]) JacksonIteratorDecoder.create().decode(response, byte[].class)).isEmpty();
   }
 }